Clean up file hasher objects
